astra-connectivity/requirements.txt
cassandra-driver
requests
python-dotenv